add_enginetype_enginelocation

+++ AutoShow/AdminCrudForm.cs
@@ -22,9 +22,20 @@ namespace AutoShow
             _context = context;
             _adminMainForm = adminMainForm;
             _option = option;
-            if(_option == Option.BodyType)
+            if (_option == Option.BodyType)
             {
-                DataGridView.DataSource = _context.BodyTypes.Select(b => new { BodyTypeName = b.BodyTypeName }).ToList();
+                HeaderLabel.Text = "Типы кузова";
+                DataGridView.DataSource = _context.BodyTypes.Select(b => new { b.BodyTypeName }).ToList();
+            }
+            else if(_option == Option.EngineLocation)
+            {
+                HeaderLabel.Text = "Распол.двиг.";
+                DataGridView.DataSource = _context.EngineLocations.Select(e => new { e.EngineLocationName }).ToList();
+            }
+            else if(_option == Option.EngineType)
+            {
+                HeaderLabel.Text = "Типы двигателя";
+                DataGridView.DataSource = _context.EngineTypes.Select(e => new { e.EngineTypeName }).ToList();
             }
         }
 
@@ -32,7 +43,15 @@ namespace AutoShow
         {
             if(_option == Option.BodyType)
             {
-                DataGridView.DataSource = _context.BodyTypes.Select(b => new { BodyTypeName = b.BodyTypeName }).ToList();
+                DataGridView.DataSource = _context.BodyTypes.Select(b => new { b.BodyTypeName }).ToList();
+            }
+            else if(_option == Option.EngineLocation)
+            {
+                DataGridView.DataSource = _context.EngineLocations.Select(e => new { e.EngineLocationName }).ToList();
+            }
+            else if(_option == Option.EngineType)
+            {
+                DataGridView.DataSource = _context.EngineTypes.Select(e => new { e.EngineTypeName }).ToList();
             }
         }
 
@@ -50,23 +69,42 @@ namespace AutoShow
 
         private void AddButton_Click(object sender, EventArgs e)
         {
-            var createUpdateAdminForm = new CreateUpdateAdminForm(_context, this, _option);
+            var createUpdateAdminForm = new CU_OneField_AdminForm(_context, this, _option);
             createUpdateAdminForm.Show();
             this.Hide();
         }
 
         private void UpdateButton_Click(object sender, EventArgs e)
-        {
+        {    
+            if (DataGridView.SelectedRows.Count < 1)
+            {
+                MessageBox.Show("Выберите кортеж");
+                return;
+            }
+            if(DataGridView.SelectedRows.Count > 1)
+            {
+                MessageBox.Show("Выбрано слишком много кортежей");
+                return;
+            }
             if (_option == Option.BodyType)
             {
-                if (DataGridView.SelectedRows.Count < 1)
-                {
-                    MessageBox.Show("Выберите кортеж");
-                    return;
-                }
                 string bodyTypeName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
                 var bodyType = _context.BodyTypes.Single(b => b.BodyTypeName == bodyTypeName);
-                var createUpdateAdminForm = new CreateUpdateAdminForm(_context, this, _option, bodyType);
+                var createUpdateAdminForm = new CU_OneField_AdminForm(_context, this, _option, bodyType);
+                createUpdateAdminForm.Show();
+            }
+            else if(_option == Option.EngineLocation)
+            {
+                string engineLocationName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                var engineLocation = _context.EngineLocations.Single(en => en.EngineLocationName == engineLocationName);
+                var createUpdateAdminForm = new CU_OneField_AdminForm(_context, this, _option, engineLocation);
+                createUpdateAdminForm.Show();
+            }
+            else if(_option == Option.EngineType)
+            {
+                string engineTypeName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                var engineType = _context.EngineTypes.Single(en => en.EngineTypeName == engineTypeName);
+                var createUpdateAdminForm = new CU_OneField_AdminForm(_context, this, _option, engineType);
                 createUpdateAdminForm.Show();
             }
             this.Hide();
@@ -80,14 +118,29 @@ namespace AutoShow
                 MessageBox.Show("Выберите кортеж");
                 return;
             }
+            if (DataGridView.SelectedRows.Count > 1)
+            {
+                MessageBox.Show("Выбрано слишком много кортежей");
+                return;
+            }
             if (_option == Option.BodyType)
             {
                 string bodyTypeName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
                 _context.BodyTypes.Remove(_context.BodyTypes.Single(b => b.BodyTypeName == bodyTypeName));
-                if (_context.SaveChanges() > 0)
-                {
-                    MessageBox.Show("Данные успешно удалены");
-                }
+            }
+            else if(_option == Option.EngineLocation)
+            {
+                string engineLocationName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                _context.EngineLocations.Remove(_context.EngineLocations.Single(en => en.EngineLocationName == engineLocationName));
+            }
+            else if(_option == Option.EngineType)
+            {
+                string engineTypeName = DataGridView[0, DataGridView.SelectedRows[0].Index].Value.ToString();
+                _context.EngineTypes.Remove(_context.EngineTypes.Single(en => en.EngineTypeName == engineTypeName));
+            }
+            if (_context.SaveChanges() > 0)
+            {
+                MessageBox.Show("Данные успешно удалены");
             }
             RefreshData();
         }
